% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmvoutlier}
\alias{rmvoutlier}
\title{Removing outliers.}
\usage{
rmvoutlier(x, coef = 2.5)
}
\arguments{
\item{x}{A vector of a variable}

\item{coef}{A number of the times of standard deviation. Default is \code{2.5}.}
}
\value{
Location of outliers in the vector
}
\description{
Function for removing outliers.
}
\examples{
data("zn")
# log-transformation
hist(zn$Zn)
zn$Zn <- log(zn$Zn)
hist(zn$Zn)
# remove outliers
k <- rmvoutlier(zn$Zn, coef = 2.5)
k

}
