% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zn}
\alias{zn}
\title{Spatial datasets of trace element Zn.}
\format{
\code{zn}: A tibble of trace element Zn with 894 rows and 12 variables
}
\usage{
zn
}
\description{
Spatial datasets of trace element Zn.
}
\author{
Yongze Song \email{yongze.song@outlook.com}
}
\keyword{dataset}
\keyword{elements}
\keyword{trace}
